Changed int to long

+++ ProjectB_Museum_DeMystery/Visitor.cs
@@ -1,12 +1,12 @@
 class Visitor
 {
-    public int Code;
+    public long Code;
     public string Name;
     public string Email;
-    public int Phonenumber;
+    public long Phonenumber;
     public GuidedTour guidedTour;
 
-    public Visitor(string name, string email, int phonenumber)
+    public Visitor(string name, string email, long phonenumber)
     {
         Name = name;
         Email = email;
